+++ app/src/androidTest/java/com/android/labellens/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.android.cameraxtest
+package com.android.labellens
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.android.cameraxtest", appContext.packageName)
+        assertEquals("com.android.labellens", appContext.packageName)
     }
 }
